Allow 200 response for endpoints in build
