+++ .changeset/cool-ducks-mix.md
@@ -1,0 +1,5 @@
+---
+'astro': patch
+---
+
+Allow 200 response for endpoints in build
