+++ packages/astro/test/fixtures/with-endpoint-routes/src/pages/invalid-redirect.json.ts
@@ -1,0 +1,11 @@
+export const get = () => {
+  return new Response(
+    undefined,
+    {
+      status: 301,
+      headers: {
+        Location: 'https://example.com',
+      }
+    }
+  );
+};
